Errs

+++ scripts/dao/dao.sh
@@ -48,8 +48,8 @@ kube::dao::gen_add_cluster_db(){
     ((host_2=total_nodes))
     for wkr in "${workers[@]}"; do
         echo "192.168.122.${host_1} ${wkr}.kubernetes.local ${wkr} 10.200.${host_2}.0/24 node" >> ${MACHINES_NEW_DB}
-        ((host_1++))
-        ((host_2++))
+        ((host_1=host_1+1))
+        ((host_2=host_2+1))
     done
 }
 kube::dao::merge_dbs(){
@@ -67,26 +67,26 @@ kube::dao::gen_cluster_db(){
     if [ "${#servers[@]}" -gt 1 ]; then
         if [ -n "$lbvip" ]; then 
             echo "192.168.122.${host_1} ${lbvip}.kubernetes.local ${lbvip} 0.0.0.0/24 lbvip" >> ${MACHINES_DB}
-            ((host_1++))
+            ((host_1=host_1+1))
         fi
         for lb in "${lbs[@]}"; do
             echo "192.168.122.${host_1} ${lb}.kubernetes.local ${lb} 0.0.0.0/24 lb" >> ${MACHINES_DB}
-            ((host_1++))
+            ((host_1=host_1+1))
         done
         for svr in "${servers[@]}"; do
             echo "192.168.122.${host_1} ${svr}.kubernetes.local ${svr} 0.0.0.0/24 server" >> ${MACHINES_DB}
-            ((host_1++))
+            ((host_1=host_1+1))
         done
     else
         svr=${servers[0]}
         echo "192.168.122.${host_1} ${svr}.kubernetes.local ${svr} 0.0.0.0/24 server" >> ${MACHINES_DB}
-        ((host_1++))
+        ((host_1=host_1+1))
     fi
     local host_2=0
     for wkr in "${workers[@]}"; do
         echo "192.168.122.${host_1} ${wkr}.kubernetes.local ${wkr} 10.200.${host_2}.0/24 node" >> ${MACHINES_DB}
-        ((host_1++))
-        ((host_2++))
+        ((host_1=host_1+1))
+        ((host_2=host_2+1))
     done
 }
 kue::dao::merge_dbs(){
@@ -148,7 +148,7 @@ kube::dao::cluster::servers(){
     echo "$(kube::dao::cluster::get_type_is "server")"
 }
 kube::dao::cluster::all(){
-    local db=$(kube::dao::cluster::curr_db $1)
+    local db=$(kube::dao::cluster::curr_db)
     cat ${db}
 }
 kube::dao::cluster::lb(){
@@ -159,17 +159,17 @@ kube::dao::cluster::lb(){
     echo "${lb}"
 }
 kube::dao::cluster::get_type_is_not(){ 
-    local db=$(kube::dao::cluster::curr_db $1)
+    local db=$(kube::dao::cluster::curr_db)
     local result=$(awk -v value="$1" '$5 != value {print $0}' ${db})
     echo "$result"
 }
 kube::dao::cluster::get_type_is(){ 
-    local db=$(kube::dao::cluster::curr_db $1)
+    local db=$(kube::dao::cluster::curr_db)
     local result=$(awk -v value="$1" '$5 == value {print $0}' ${db})
     echo "$result"
 }
 kube::dao::cluster::get(){
-    local db=$(kube::dao::cluster::curr_db $1)
+    local db=$(kube::dao::cluster::curr_db)
     local values=($(awk -v value="$1" -v col="$2"  '$5 == value {print $col}' ${db}))
     for e in "${values[@]}"; do
         echo "$e"
@@ -180,7 +180,7 @@ kube::dao::cluster::curr_nodes(){
     echo "$result"
 }
 kube::dao::cluster::count(){
-    local db=$(kube::dao::cluster::curr_db $1)
+    local db=$(kube::dao::cluster::curr_db)
     local value=$(awk -v value="$1" '$5 == value {count++} END {print count}' ${db})
     echo "$value"
 }
@@ -202,7 +202,9 @@ kube::dao::cluster::lock(){
     mv ${MACHINES_DB} ${MACHINES_DB_LOCK}
 }
 kube::dao::cluster::unlock(){
-    mv ${MACHINES_DB_LOCK} ${MACHINES_DB} 
+    if [ -f "${MACHINES_DB_LOCK}" ]; then 
+        mv ${MACHINES_DB_LOCK} ${MACHINES_DB} 
+    fi
 }
 kube::dao::cluster::is_locked(){
     if [ -f ${MACHINES_DB_LOCK} ]; then
